update onerun.sh script

+++ onerun/onerun.sh
@@ -1,7 +1,12 @@
 #!/bin/sh
 
-sudo apt-get update
-sudo apt-get install openjdk-11-jdk
+if which java > /dev/null 2>&1; 
+then
+    sudo apt-get update
+	sudo apt-get install openjdk-11-jdk
+else
+    echo "Java not installed"
+fi
 
 export TOMCATPATH=$HOME/tomcat9
 
@@ -14,9 +19,13 @@ fi
 
 FILENAME=$TOMCATPATH'.tar.gz'
 
-wget -O $FILENAME https://dlcdn.apache.org/tomcat/tomcat-9/v9.0.75/bin/apache-tomcat-9.0.75.tar.gz
-
-tar -xf $FILENAME -C $TOMCATPATH --strip-components=1
+if [ -f $FILENAME]
+then
+	echo "File exists!"
+else
+	wget -O $FILENAME https://dlcdn.apache.org/tomcat/tomcat-9/v9.0.75/bin/apache-tomcat-9.0.75.tar.gz
+	tar -xf $FILENAME -C $TOMCATPATH --strip-components=1
+fi
 
 TOMCATUSERFILE=$TOMCATPATH/conf/tomcat-users.xml
 CONTEXTFILE=$TOMCATPATH/webapps/manager/META-INF/context.xml
@@ -28,5 +37,14 @@ chown --reference=$CHMODCP $TOMCATUSERFILE
 cp $HOME/git/onerun/onerun/context.xml $CONTEXTFILE
 chown --reference=$CHMODCP $CONTEXTFILE 
 
+WEBAPPSPATH=$TOMCATPATH/webapps/
+
+if [ -d $TOMCATPATH/webapps/]
+then
+	JENKINNAME=$WEBAPPSPATH/jenkins.war
+	wget -O $JENKINNAME https://get.jenkins.io/war-stable/2.401.1/jenkins.war
+else
+	echo "No directory exists!"
+fi
 
 $TOMCATPATH/bin/startup.sh
